Add logos-codegen CLI
