+++ Cargo.toml
@@ -1,9 +1,5 @@
 [workspace]
-members = [
-    "logos",
-    "logos-derive",
-    "tests",
-]
+members = ["logos", "logos-cli", "logos-codegen", "logos-derive", "tests"]
 
 [profile]
 release = { lto = true }
