+++ logos-codegen/src/generator/tables.rs
