+++ logos-codegen/src/graph/impls.rs
